+++ src/main/java/org/secureauth/sarestapi/data/BaseResponse.java
@@ -1,0 +1,31 @@
+package org.secureauth.sarestapi.data;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+import javax.xml.bind.annotation.XmlRootElement;
+
+/**
+ * Created by rrowcliffe on 5/8/16.
+ */
+@XmlRootElement
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class BaseResponse {
+    private String status;
+    private String message;
+
+    public String getStatus() {
+        return status;
+    }
+
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
